Домашна 4/sentiment-service/sentiment-analysis/requirements.txt
pandas
numpy
transformers
torch
scikit-learn
tqdm
psycopg2-binary
python-dotenv
dateparser
bs4
selenium
webdriver_manager
transformers
scikit-learn
torch
psycopg2